refactor(Event): Move Event class to new event.py
Moved the Event class to it's own submodule within io.
